Change anchors to buttons
This switches a ton of anchor tags for button tags,
thereby removing a ton of ESLint warnings and
hopefully preparing the Admin UI for future
accessibility updates.

Since this is supposed to be just a semantic update,
everything should look the same as before, particularly
every clickable thing.

Does not fix the warnings in TimeSeriesStatistics.js because
I don't currently know how to display that.