Fix exception when opening timepicker
If you open a time picker (e.g. "startDate" in the event
details), the ui crashes with an exception:
"Format string contains an unescaped latin alphabet
character `n` ".

This fixes that by downgrading the date-fns dependency.
Allegedly the issues is with material-ui not providing
the correct types for date-fns, but downgrading material-ui
would cause more problems.