Refresh workflow operations in event details
In the old admin ui, the status of workflow operations in the event
details was updated in real-time, allowing you to "watch" your
workflow running. This updates our admin ui to do the same by
re-fetching workflow operations every 5 seconds.

Unfortunately, this is causing rerenders even if the response
from the Opencast backend has not changed between calls.
We could probably memoize this somehow, but I can't find a
straightforward way, at least not one that is more straightfoward
than switching to redux toolkit (like in #214).