Show conflicting scheduling events for new events
When trying to schedule a new event, there already is a notification
if the new event would cause a conflict with an already scheduled
event. However, it would not tell you with which event, leaving you
guessing on how to resolve the conflict. This display the conflicting
event, same as when trying to reschedule an event in the event
details.