Improve new event stepper buttons
Clicking on the circles in the top part of the new event dialog
only works sometimes. This commit makes it more consistent,
by removing some validation logic where it really does not belong.
This should allow users to click on the circles in the top to go to
previously completed steps as well as the current step. However,
the old ui, it does not allow users to click on the circle of the next
step to validate the current step.