Don't render "invalid date" in table cells
The constructor for javascript `Date`s will return "invalid date" for
invalid date strings. This does not look very nice in our tables,
so this changes it to render an empty string instead. Arguably this
is fine because it is not the responsibility of the view to make sure
the dates it gets are correct.