Remove redux persist
Redux persist in its current form is causing to many
problems to keep it around. Therefore this simply
removes it.

The most common problem with redux persist is
that users make changes to configuration, which
are then not taken over into the ui as they are blocked
by the existing redux state. This can also happen
when new versions of the ui are released. Resolving
this requires a hard page reload which is simply un-
userfriendly.

Whatever efficiency benefits redux persist may bring,
they are not worth requiring clearing your browser
cache every other day.