Add quickstart guide and update development instructions
Based on the discussion in the technical meeting today,
this adds a quick start guide for people who want
to copy and paste commands into their console.
It also changes around some stuff in the development
section to make clear the the setup with the proxy
server is considered best practice.