Update Material UI to v5
Updates material-ui dependencies from version 4 to
version 5. This affects date/time-pickers for e.g.
scheduling or changing metadata dates, as well
as the circles at the top of the new event/series
wizards.

This does change how the date/time-pickers
look, because we were using the default style
before which now changed. In my opinion this
is for the better.

Also adds back the ability to type in the date field
instead of being forced into using a picker, although
that does seem to cause errors. Will investigate.