Fix unique key issue in WizzardStepper
Fixes a web console complaint about missing unique keys in the
WizzardSteppers. I deem using .filter() a nicer solution than a ternary
operator that checks if the hidden property of the current label object
is false and rendering an empty fragment. This is probably also what
made the console unhappy.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>