Remove the proxy server
Well, the outer one at least. For context: This app had two proxies:

One comes with `create-react-app` as a middleware in the dev server,
proxying all the requests the frontend makes (except for the static
JS, CSS, etc. assets this project is generating) to some other
server.

In production, Opencast delivers the admin UI assets and the
backend under the same URL, but in development this is not true.
The development server delivers the assets and the backend is ...
well it's complicated.

So far you either started a static file server which was acting as a
mock backend, or you started **another** proxy server that piped the
requests through to an Opencast instance of your choosing while
tacking on some digest authentication.

Well, the dev server proxy can't (easily) do digest auth, but the fact
that we insist on digest auth is a historical artifact, anyway, at
least as far as I know. It can do basic auth, though, and it **is** a
proxy, so it can do everything we need!

This also updates the `README` appropriately and cleans it up some in
general (though not completely).