Pull the static file server into the CRA proxy as well
With this, the outer npm-project is finally obsolete.
See the updated `README` for how to use it and the proxy server.
I abstained from restructuring the project any further
to keep the PR focused, but a PR that will do nothing
but move things around is in the pipeline.

Note, this is based on #394.