Persist table pagination settings
Persists changes to table pagination information, most notably
the setting for how many items per page are displayed.

Since this information is stored and applied to ALL tables, setting
the limit from 10 to 20 for the events table will also set it to 20
for the series table and all other tables. If we say that is no good,
and we want the settings to be per table instead, I'll have to create
a different, more extensive PR.