Fix date table filter from event column
In the event table you can click on a date in the
date column to set a filter. That filter was not very
helpful, as it would only show the event of the row
you clicked on. Now it shows all events of that day,
bringing it in line with how setting a date filter
manually works.