Don't show notification for tobira server error
The notification for a server error when fetching
tobira information from Opencast was already
shown on the ACL tab, before the Tobira tab. This
was especially annoying if you had not intention
of using Tobira, as the notification would show up
anyway. This disables the notification and instead
puts some information in the logs. A better, but
currently not easily possible solution would be
to confine the notification to the tobira tab.