Fix required metadata not being required
In the backend, metadata catalogs can be
configured as needed. One thing that can be
configured for each metadata field is if is
considered required. A user should not be able
to leave a required field empty (or set it to empty).
Our validation was ignoring said configuration,
merely enforcing the title field.

This fixes the issue by dynamically creating a validation
schema based on configuration for the
existing metadata fields. Although it really
feels like Yup does not want me to do this.