Display active workflow warning
An event with an active workflow cannot be edited.
However, the warning informing the user about that
was only shown on the access tab of the event
details. This aims at showing the warning no
matter what tab the user is on.

Does not work on the main metadata tab for
reasons yet unknown to me.