Properly set theme default configuration
Instead of properly setting the default value for `admin.themes.enabled`
in the admin interface, we temporarily set it in the Opencast
configuration. That's not a great workaround and it has a few problems,
but I let us backport the admin interface to `r/15.x` to fix the major
problems there.

Not that 15 is our legacy release and we already have a two 16.x
versions out, it's time to properly set the default value.