Fix the URL for the "about" configuration in the admin UI
The admin UI can fetch HTML files from the `ui-config` module
to display in its "about" view which you can reach by the links
in the footer. (If you have that feature enabled.)

The URL used to fetch these was a relative one, which works
in "dev mode" since the site is served from the root,
but it doesn't work in production, where the link is resolved
relative to `/admin-ui`.