Include admin interface version
This patch adds the admin interface build date and the git commit it was
built from to the top of the HTML head section. This makes it easy for
developers to check which version of the admin interface exactly they
are dealing with.

This is hidden from regular users and therefor shouldn't make them
confuse this with the version of Opencast itself.