Fix new filters not being loaded
When a new table filter was added to Opencast,
it would not show up in the filters dropdown
unless the page cache was cleared. This should
fix that.

I legit do not understand why the code was the
way it was before this commit. If a reviewer
could enlighten me why the previous code
makes sense that would be great, but personally
I see no harm in fetching the respective table
filters everytime we load a new table.