Fix test containers
This patch fixes the test containers, which were failing since `git` is
now required for building/running the admin interface.

If `git` is not installed, the tooling will now read the `commit` file if
present (we create one during the container build on GitHub Actions). If
the file also does not exists, the build continues with an empty commit
string.