Render capture agent input names more nicely
Currently were displaying the whole translation
string for a capture agent input if a translation
could not be found, which is rather likely. This
commit display the input id in case there is no
translation, which should be a reasonable value
in most cases. (This is also how it worked in the
old admin ui)