Remember table sorting
Tables can be sorted by columns, ascending or
descending. However, when switching to a
different tab and back again, the sorting reverts
to the default. With this change, the sorting is
remembered.

Does not persist the sorting, meaning it still
reverts to the default after a page reload.
I did not do this because the only to way to
get rid of persisted sorting would be to the
clear the browsers cache, and I did not want
alienate people.

Also sneakily changes the default sorting
of the series table in the same way as #1008.