Reduce duplicate code for NavBar components
This patch attempts to reduce duplicate code by
moving components that would usually be
passed to the NavBar component INTO the
NavBar component.

This does come with the trade-off that the
NavBar component is now less general. However,
as it is the NavBar component was pretty
pointless anyway so this should not be an issue.