Show username in user role dropdown
The options in the dropdown for selecting a user in the
access policy tab would show the user role
(aka the thing actually being selected) instead of
the user information like in the input field.
This fixes that.

Also adds some sensible fallbacks in case someone
decides to have users without names.