Fix Can't create theme when uploading a file
When trying to create a theme, you can upload
files. After the upload completes, you should be
able to proceed to the next step. However, the
"Next" button remained disabled, making it
impossible to proceed.
(Well, not actually impossible, going a step back
and then forward again would fix the issue, but w/e).

This fixes that, but crudely. We may want to rewrite
the theme dialog because uploading each file individually
and forcing the user to wait on that seems pretty
bad UX to me.