Reduce duplicate code - button like anchor
We have many buttons using the "button-like-anchor"
class. This patch introduces a button component
for these buttons, with built-in tooltip and access
check to better support common code combinations
for those buttons.
In and of itself not a particularly impressive
component, but the underlying goal here is to
push the app more towards using common basic
components.