Improve UX of series path editing
Previously, clicking the "edit path" button
of a series in the Tobira tab would open the
"select target page" menu showing the currently
selected path. This changes it to instead open
the menu showing the parent path instead.

In addition, this will also mark the currently
selected page with a checkmark, as long as no
other page has been selected.