Remove custom filtering for dropdowns
This should make the search for searchable
dropdowns work as expected, meaning the
search string is now matched correctly against
the existing labels. The custom filtering this patch
removes only serves to break the search, and what
it tries to achieve the 'react-select' search already
does out of the box.