Get rid of `styled-components`
We were using it in one place only,
and while there are certainly some advantages over dynamic inline styles,
we are using those pervasively as well anyway.
Plus, we pull in MUI for some stuff,
which by default relies on Emotion for that stuff,
so if we want those benefits,
I would propose that it's at least worth considering
to get them through that.