Explicitly `npm install @emotion/react @emotion/styled`
As per https://mui.com/material-ui/getting-started/installation/#default-installation.

These were installed as (optional?) transitive (peer?) dependencies,
but that can of course change randomly.
I've seen this happen while preparing a custom build of this repo.

Note that there is also https://mui.com/material-ui/getting-started/installation/#with-styled-components,
which is arguably more fitting to this project,
because we were already using `styled-components`,
but I opted for this solution because it seemed simpler and better supported,
and also because our usage of `styled-components` isn't really justified.
However, see the previous commit for that.