Enforce event ordering
**WILL BREAK THE UI WITHOUT THE REQUIRED BACKEND CHANGES:** https://github.com/opencast/opencast/pull/6461

The events table is sorted by a sort parameter. By default, this is
the start date. If there are multiple events with the same start date,
the sorting for those is left to the elasticseach index in the
backend. Elasticsearch however does not guarantee any ordering.
This may result in the order of events changing unexpectedly like
in #1102.

This patch aims to solve this by adding a secondary
sort parameter to the
events query. The primary parameter still takes priority.
For the secondary parameter, `uid` was chosen as it is the only
field guaranteed to be unique.

More sorting parameters can lead to increased query times.
If someone with *many* events could test
the impact of these changes
that would be great.