Fix "Typing in date input ends early" (#1084)
Trying to type in a date in for example the start date table filter
is not working. The input is interpreted as a date
immediately. For example, typing "1" will result in
"01/01/2025", even if you wanted to type something
completely different.

This patch fixes that by enabling strict parsing. Arguably
not the best solution, but it works for now.