Split presenters at ";"
It was brought to my attention that this was a feature in the
previous admin ui,
so this patch adds it back in.

When adding multiple presenters to an event in the
metadata tab,
you can now add multiple presenters at the same time by delimiting
them with a ";". For example, writing "Pacman; Duck Hunt" would result
in both "Pacman" and "Duck Hunt" being added as presenters.
Will also work for other metadata fields that work the same way,
e.g. contributers.