Remove btm spacer, fix scroll bars in modals (#1053)
Since we decided that our modals should be
scrollable, some of them always have a
scrollbar showing, due to an absolutely
positioned white box at the very bottom of
the modal. I am not sure what the original
purpose of this "btm-spacer" component
was supposed to be, but it seems superfluous
to me and so I am removing it. This should stop
the scrollbars from appearing unnecessarily.