Don't fetch page info in MainNav component
The main navigation component fetches information
for the page it is loading. As far as I understand it,
the idea is to prefetch data to have it ready
on page load. This makes pages load slower. Furthermore, the page
itself still has to fetch that information anyway,
because it might be loaded by other means, meaning
we are doing the same (potentially expensive)
backend requests twice. Therefore, this patch
removes all backend requests from the main navigation component.