Make events table load faster
When fetching events, we would then make subsequent backend calls and
await on each of them before returning. This would cause the response
time to be several seconds instead of milliseconds even for a low number
of events. This moves the subsequent backend calls into the component
that actually requires the information so it can make the calls itself
without delaying the table loading.