Fix hiding tobira tab in series create modal
In the series create modal, we query ´/admin-ng/series/new/tobira/page`
to check if Tobira is available. However, if the user does not have the
role required for that endpoint (ROLE_UI_SERIES_DETAILS_TOBIRA_EDIT),
the endpoint will return an error html page which we are not expecting,
causing the ui to crash.
This fixes the problem by checking if the user has the required role to
make the request to the endpoint. If the user does not have the role,
the tobira tab is then not shown.