Rewrite metadata fields for accessibility
The goal here is to make our metadata fields more accessible and
easier to work with in general.
Instead of rendering different components based on state,
the same component is always rendered and only the CSS changes based on
whether the component is focused. This keeps or DOM consistent which in
turn makes it easier to introduce keyboard
navigation between fields or focus trapping for modals.