Typecheck translation strings
If we attempt to translate a string which does not exist in our
translations JSON, we now get a build error to warn us of that.
This is intended to help avoid mistakes like in #1145,
where a translation string was simply misspelled.
We do so by leveraging the typescript capabilities of i18n.

Unfortunately, we also get a bunch of strings from the backend
which can be translation strings, any other string or possibly either!
Which is why you will see a lot "as ParseKeys" typecasts.
We could theoretically check those stringsat runtime (if we somehow
knew that they must be translation strings) and render some default
message, but I personally think it is better to then render a "wrong"
translation string, because it may still contain valuable information,
even if that information is not rendered nicely. For example,
"EVENTS.STATUS.COMPLETED" can still signal that event processing
has finished, even if it is a bit hard to read and requires
english skills.