Fix typescript conflict for initial state of table field "resource"
Fixes #1149.
Fixes a typescript error that was caused by adding the empty string
to a string enumeration.

This patch fixes that removing the empty string from the type
and fixing reason the empty string was introduced in the first place
in a different manner (also see commit 28aa9821aac65f8acea9fd2555d87125bc969ed3).

This has the downside of forcing us to set "resource" to a concrete
value, which is technically wrong.
However, I find this preferable to allowing "resource" to be undefined.
It makes our typing tighter and easier to handle. Furthermore,
"resource" must be set on table load anyway, so the technically wrong
value gets overwritten asap.