Fix end time date display for scheduling
Fixes display behaviour for the end time date display on the scheduling
pages to go back to how it was before.
Namely the end time date is now again displayed in the same row
as the end time, and is only actually displayed if the end time is
"before" the start time (at which the ui assumes that the end time has
to be on the next day).