Make wizard stepping more robust
Changes the code surrounding the wizard stepper (the circle at the top
of e.g. the create event modal) to rely on string literals instead
of indices. This should make it harder to accidentally break anything
when making changes to the code.

Also fixes web console errors pertaining to "React.Fragment" in the
WizardStepper component (did not cause any functional errors).

Also fixes #1098.