Improve validation/confirmation of selected page
The validation logic does not consider the fact that
pressing the button to add a new subpage (which
will open the input fields for title and path segment)
already selects the page for formik.
Since we added checkmarks which serve that purpose now,
the former selction process is removed with this commit,
allowing for better, or at least easier validation (well
maybe not even easier, but at least I won't have to redo
the validation logic itself).